Move `external-tracer` from `bus-mapping` to a new crate only for test.
